refactor: 패키지 세분화 및 파일 배치 변경
