+++ src/main/java/domain/card/shufflestrategy/ShuffleStrategy.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.card.shufflestrategy;
 
+import domain.card.Card;
 import java.util.List;
 
 public interface ShuffleStrategy {
