+++ src/main/java/domain/match/MatchResult.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.match;
 
 public enum MatchResult {
