+++ src/test/java/domain/participant/bet/BetAmountTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant.bet;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
